+++ models/Product.cs
@@ -5,7 +5,7 @@ public class Produto
     public string CategoriaNome { get; set; }
     public string Nome { get; set; }
     public string SKU { get; set; }
-    public decimal EstoqueAtual { get; set; }
+    public int EstoqueAtual { get; set; } // old: decimal
     public decimal Preco { get; set; }
     public bool Ativo { get; set; }
     public DateTime DataCadastro { get; set; }
